+++ cr.py
@@ -1,2 +1,0 @@
-import os  
-f=open('core/application/behaviors/react/utils/__init__.py','w',encoding='utf-8')